Separation of jvm-specific code from ConnectionTest
